Updates to Nipypeline, currently tested until step 6 (Step 7 still running and seems to be veeery slow) Additionally added a resampled+padded segmentation node for umc_tse_native.
Remember it still only runs if you update your c3.py file in Nipype interfaces!
Also, it is not tested with the actual templates, but dummy templates! The path to git lib needs to include the template for this to work out of the box, so currently we have just placed a dummy file locally.